Enemy Spawn Locations
Enemy spawn locations have a trigger box and a spawn area. The amount of
enemies spawned and the amount of times the spawn will trigger can each
be adjusted with a variable.
What type of Enemy is spawned can be decided by an Enum.
Currently completed in blueprint will be converted to c++ at a later
time